hacked around weird testing issue
